+++ teachers/api/urls.py
@@ -11,6 +11,8 @@ urlpatterns = [
     path('classroom/joined/', views.student_joined_classes_view),
     path('classroom/detail/<int:classroom_id>/', views.classroom_detail_view),
     path('classroom/students/<int:classroom_id>/', views.classroom_students_view),
-    path('classroom/attach-deck/<int:classroom_id>/', views.attach_deck_view),
+    path('classroom/attach-deck/<int:classroom_id>/', views.teacher_attach_deck_view),
     path('classroom/<int:classroom_id>/student/<int:student_id>/stats/', views.student_statistics_view),
+    path('classroom/students/attach-deck/<int:classroom_id>/', views.student_attach_deck_view),
+    path('classroom/<int:classroom_id>/student/<int:student_id>/attached-deck/', views.student_get_attached_deck_view),
 ]